Move Frida config out of Frameworks
The version of Sideloader that atvloadly uses attempts to sign
everything in the Frameworks directory, including the Frida config.
This leads to an error because the Frida config is not a Mach-O file.

See https://reddit.com/r/sideloaded/comments/1lgwohj/f/mz4w3i8/.

To fix, store the Frida config in the parent directory instead as Frida
supports this too.